Move utility method from LedgerMetadata to test
The method is only used in the test, so it should live there.

Master issue: #281
